Proof of concept: Fast search for first potential byte.
This adds an optimization to search for the first potential
byte using char classes. This optimization is inspired
by an existing optimization that searches for the first
potential byte (or range of bytes) when the pattern
starts with one or more literals.

This table shows the results of BenchmarkSubMatch,
which is the number of ms required to match a one MB
file looking for `909` at the end of the file.

The amount of noise indicates the number of `9` characters
in the file. As the noise increases, there are more states
in the NFA to explore, hence the search becomes slower.

 milliseconds to find a match at the end of a
 one megabyte file:

		\d{3}				9\d{2}
Noise (%)	unoptimized	optimized	leading literal
0		29.5		0.3		0.3
5		31.1		3.6		3.6
10		30.7		7.5		7.4